rename file and fix index.ts
